+++ GAN_4D.py
@@ -24,6 +24,7 @@ import glob
 import time
 import shutil
 import os
+import random
 
 from pickle import load
 from sklearn.preprocessing import QuantileTransformer
@@ -63,6 +64,9 @@ def split_tensor(index, x):
 
 print(tf.__version__)
 
+# mode = "long test"
+mode = "ROC_testing"
+
 # working_directory = '/Users/am13743/Aux_GAN_thesis/THESIS_ITERATION/TRAINING/'
 # training_directory = '/Users/am13743/Aux_GAN_thesis/THESIS_ITERATION/DATA/'
 # transformer_directory = '/Users/am13743/Aux_GAN_thesis/THESIS_ITERATION/TRANSFORMERS/'
@@ -80,7 +84,7 @@ transformer_directory = '/mnt/storage/scratch/am13743/AUX_GAN_THESIS/THESIS_ITER
 pre_trained_directory = '/mnt/storage/scratch/am13743/AUX_GAN_THESIS/THESIS_ITERATION/PRE_TRAIN/'
 training_name = 'relu*.npy'
 testing_name = 'test_relu*.npy'
-saving_directory = 'GAN_4D'
+saving_directory = 'GAN_4D_ROC'
 save_interval = 25000
 weight_of_reco_kicks_in_at = 0
 weight_of_reco_maxes_at = 0 # if weight_of_reco_maxes_at > 0:
@@ -91,7 +95,7 @@ weight_of_reco_maxes_at = 0 # if weight_of_reco_maxes_at > 0:
 # pre_trained_directory = '/hdfs/user/am13743/THESIS/PRE_TRAIN/'
 # training_name = 'relu*.npy'
 # testing_name = 'test_relu*.npy'
-# saving_directory = 'GAN_4D'
+# saving_directory = 'GAN_4D_ROC'
 # save_interval = 25000
 # weight_of_reco_kicks_in_at = 0
 # weight_of_reco_maxes_at = 0 # if weight_of_reco_maxes_at > 0:
@@ -100,8 +104,15 @@ weight_of_reco_maxes_at = 0 # if weight_of_reco_maxes_at > 0:
 
 calculate_ROC = True
 test_boosting = False
-
+quit_at_iteration = int(1E30)
 batch_size = 50
+virtual_batch_size = 50
+
+FoM_ID = np.random.randint(1000000,9999999)
+if mode == "ROC_testing":
+	calculate_ROC = True
+	save_interval = 5000
+	quit_at_iteration = 70002
 
 # G_architecture = [1000,1000]
 # D_architecture = [1000,1000]
@@ -125,18 +136,19 @@ list_of_training_files = glob.glob('%s%s'%(training_directory,training_name))
 # 		os.remove(file_i)
 # except:
 # 	print('/CORRELATIONS/ already clean')
-try:
-	files_to_remove = glob.glob('%s%s/*.png'%(working_directory,saving_directory))
-	for file_i in files_to_remove:
-		os.remove(file_i)
-	files_to_remove = glob.glob('%s%s/*.h5'%(working_directory,saving_directory))
-	for file_i in files_to_remove:
-		os.remove(file_i)
-	files_to_remove = glob.glob('%s%s/*.npy'%(working_directory,saving_directory))
-	for file_i in files_to_remove:
-		os.remove(file_i)
-except:
-	print('Output directory already clean')
+if mode != "ROC_testing":
+	try:
+		files_to_remove = glob.glob('%s%s/*.png'%(working_directory,saving_directory))
+		for file_i in files_to_remove:
+			os.remove(file_i)
+		files_to_remove = glob.glob('%s%s/*.h5'%(working_directory,saving_directory))
+		for file_i in files_to_remove:
+			os.remove(file_i)
+		files_to_remove = glob.glob('%s%s/*.npy'%(working_directory,saving_directory))
+		for file_i in files_to_remove:
+			os.remove(file_i)
+	except:
+		print('Output directory already clean')
 
 
 print(' ')
@@ -145,9 +157,6 @@ print(' ')
 
 
 optimizer = tf.keras.optimizers.Adam(lr=0.0004, beta_1=0.5, decay=0, amsgrad=True)
-
-kernel_initializer_choice='random_uniform'
-bias_initializer_choice='random_uniform'
 
 ##############################################################################################################
 # Build Generative model ...
@@ -257,7 +266,7 @@ H = Flatten()(d_input)
 
 for layer in D_architecture:
 
-	H = Dense(int(layer),kernel_initializer=kernel_initializer_choice,bias_initializer=bias_initializer_choice)(H)
+	H = Dense(int(layer))(H)
 	H = LeakyReLU(alpha=0.2)(H)
 	H = Dropout(0.2)(H)
 d_output = Dense(1, activation='sigmoid')(H)
@@ -341,9 +350,13 @@ training_time = 0
 
 t0 = time.time()
 
+random.shuffle(list_of_training_files)
+
 for epoch in range(int(1E30)):
 
 	for file in list_of_training_files:
+
+		batch_size = virtual_batch_size
 
 		print('Loading initial training file:',file,'...')
 
@@ -353,20 +366,21 @@ for epoch in range(int(1E30)):
 
 		X_train[:,1:7] = (X_train[:,1:7] * 2.) - 1.
 
-		if iteration == -1:
-			plt.figure(figsize=(5*4, 3*4))
-			subplot=0
-			for i in range(0, 6):
-				for j in range(i+1, 6):
-					subplot += 1
-					plt.subplot(3,5,subplot)
-					if subplot == 3: plt.title(iteration)
-					plt.hist2d(X_train[:,i+1], X_train[:,j+1], bins=50,range=[[-1,1],[-1,1]], norm=LogNorm(), cmap=cmp_root)
-					plt.xlabel(axis_titles[i])
-					plt.ylabel(axis_titles[j])
-			plt.subplots_adjust(wspace=0.3, hspace=0.3)
-			plt.savefig('%s%s/CORRELATIONS_TRAIN.png'%(working_directory,saving_directory),bbox_inches='tight')
-			plt.close('all')
+		if mode != "ROC_testing":
+			if iteration == -1:
+				plt.figure(figsize=(5*4, 3*4))
+				subplot=0
+				for i in range(0, 6):
+					for j in range(i+1, 6):
+						subplot += 1
+						plt.subplot(3,5,subplot)
+						if subplot == 3: plt.title(iteration)
+						plt.hist2d(X_train[:,i+1], X_train[:,j+1], bins=50,range=[[-1,1],[-1,1]], norm=LogNorm(), cmap=cmp_root)
+						plt.xlabel(axis_titles[i])
+						plt.ylabel(axis_titles[j])
+				plt.subplots_adjust(wspace=0.3, hspace=0.3)
+				plt.savefig('%s%s/CORRELATIONS_TRAIN.png'%(working_directory,saving_directory),bbox_inches='tight')
+				plt.close('all')
 
 		print('Train images shape -',np.shape(X_train))
 
@@ -380,11 +394,16 @@ for epoch in range(int(1E30)):
 
 		for images_for_batch in train_dataset:
 
-			if iteration % 250 == 0: print('Iteration:',iteration)
+			if mode != "ROC_testing":
+				if iteration % 250 == 0: print('Iteration:',iteration)
 
-			if iteration > 50000 and iteration % 1000 == 0: batch_size += 1
+			if iteration > 50000 and iteration % 1000 == 0: virtual_batch_size += 1
 
 			iteration += 1
+
+			if mode == "ROC_testing":
+				if iteration == quit_at_iteration:
+					quit()
 
 			if weight_of_reco_maxes_at > 0:
 				if iteration < weight_of_reco_kicks_in_at:
@@ -446,40 +465,7 @@ for epoch in range(int(1E30)):
 				images = images[:,1:]
 				samples = X_train[:,0,1:-5].copy()
 
-				plt.figure(figsize=(5*4, 3*4))
-				subplot=0
-				for i in range(0, 6):
-					for j in range(i+1, 6):
-						subplot += 1
-						plt.subplot(3,5,subplot)
-						if subplot == 3: plt.title(iteration)
-						plt.hist2d(images[:noise_size,i], images[:noise_size,j], bins=50,range=[[-1,1],[-1,1]], norm=LogNorm(), cmap=cmp_root)
-						plt.xlabel(axis_titles[i])
-						plt.ylabel(axis_titles[j])
-				plt.subplots_adjust(wspace=0.3, hspace=0.3)
-				plt.savefig('%s%s/CORRELATIONS.png'%(working_directory,saving_directory),bbox_inches='tight')
-				plt.close('all')
-
-				plt.figure(figsize=(3*4, 2*4))
-				subplot=0
-				for i in range(0, 6):
-					subplot += 1
-					plt.subplot(2,3,subplot)
-					if subplot == 2: plt.title(iteration)
-					plt.hist([samples[:noise_size,i], images[:noise_size,i]], bins=50,range=[-1,1], label=['Train','GEN'],histtype='step')
-					plt.yscale('log')
-					plt.xlabel(axis_titles[i])
-					if axis_titles[i] == 'StartZ': plt.legend()
-				plt.subplots_adjust(wspace=0.3, hspace=0.3)
-				plt.savefig('%s%s/VALUES.png'%(working_directory,saving_directory),bbox_inches='tight')
-				plt.close('all')
-
-				if test_boosting == True:
-
-					aux_gan[:,3] = aux_gan[:,3]*2.5
-					images_wide = np.squeeze(generator.predict([np.expand_dims(gen_noise,1), np.expand_dims(aux_gan,1), charge_gan, theta_gan, theta_pt_gan]))
-					images_wide = images_wide[:,1:]
-
+				if mode != "ROC_testing":
 					plt.figure(figsize=(5*4, 3*4))
 					subplot=0
 					for i in range(0, 6):
@@ -487,18 +473,54 @@ for epoch in range(int(1E30)):
 							subplot += 1
 							plt.subplot(3,5,subplot)
 							if subplot == 3: plt.title(iteration)
-							plt.hist2d(images_wide[:noise_size,i], images_wide[:noise_size,j], bins=50,range=[[-1,1],[-1,1]], norm=LogNorm(), cmap=cmp_root)
+							plt.hist2d(images[:noise_size,i], images[:noise_size,j], bins=50,range=[[-1,1],[-1,1]], norm=LogNorm(), cmap=cmp_root)
 							plt.xlabel(axis_titles[i])
 							plt.ylabel(axis_titles[j])
 					plt.subplots_adjust(wspace=0.3, hspace=0.3)
-					plt.savefig('%s%s/CORRELATIONS_2_5.png'%(working_directory,saving_directory),bbox_inches='tight')
+					plt.savefig('%s%s/CORRELATIONS.png'%(working_directory,saving_directory),bbox_inches='tight')
 					plt.close('all')
+
+					plt.figure(figsize=(3*4, 2*4))
+					subplot=0
+					for i in range(0, 6):
+						subplot += 1
+						plt.subplot(2,3,subplot)
+						if subplot == 2: plt.title(iteration)
+						plt.hist([samples[:noise_size,i], images[:noise_size,i]], bins=50,range=[-1,1], label=['Train','GEN'],histtype='step')
+						plt.yscale('log')
+						plt.xlabel(axis_titles[i])
+						if axis_titles[i] == 'StartZ': plt.legend()
+					plt.subplots_adjust(wspace=0.3, hspace=0.3)
+					plt.savefig('%s%s/VALUES.png'%(working_directory,saving_directory),bbox_inches='tight')
+					plt.close('all')
+
+				if mode != "ROC_testing":
+					if test_boosting == True:
+
+						aux_gan[:,3] = aux_gan[:,3]*2.5
+						images_wide = np.squeeze(generator.predict([np.expand_dims(gen_noise,1), np.expand_dims(aux_gan,1), charge_gan, theta_gan, theta_pt_gan]))
+						images_wide = images_wide[:,1:]
+
+						plt.figure(figsize=(5*4, 3*4))
+						subplot=0
+						for i in range(0, 6):
+							for j in range(i+1, 6):
+								subplot += 1
+								plt.subplot(3,5,subplot)
+								if subplot == 3: plt.title(iteration)
+								plt.hist2d(images_wide[:noise_size,i], images_wide[:noise_size,j], bins=50,range=[[-1,1],[-1,1]], norm=LogNorm(), cmap=cmp_root)
+								plt.xlabel(axis_titles[i])
+								plt.ylabel(axis_titles[j])
+						plt.subplots_adjust(wspace=0.3, hspace=0.3)
+						plt.savefig('%s%s/CORRELATIONS_2_5.png'%(working_directory,saving_directory),bbox_inches='tight')
+						plt.close('all')
 
 				if iteration > 0 and calculate_ROC == True:
 
-					generator.save('%s%s/generator.h5'%(working_directory,saving_directory))
-					discriminator.save('%s%s/discriminator.h5'%(working_directory,saving_directory))
-					discriminator.save_weights('%s%s/discriminator_weights.h5'%(working_directory,saving_directory))
+					if mode != "ROC_testing":
+						generator.save('%s%s/generator.h5'%(working_directory,saving_directory))
+						discriminator.save('%s%s/discriminator.h5'%(working_directory,saving_directory))
+						discriminator.save_weights('%s%s/discriminator_weights.h5'%(working_directory,saving_directory))
 
 					###################################################
 
@@ -531,11 +553,12 @@ for epoch in range(int(1E30)):
 
 					out_fake = clf.predict_proba(fake_test_data)
 
-					plt.hist([out_real[:,1],out_fake[:,1]], bins = 100,label=['real','gen'], histtype='step')
-					plt.xlabel('Output of BDT')
-					plt.legend(loc='upper right')
-					plt.savefig('%s%s/BDT_out.png'%(working_directory,saving_directory), bbox_inches='tight')
-					plt.close('all')
+					if mode != "ROC_testing":
+						plt.hist([out_real[:,1],out_fake[:,1]], bins = 100,label=['real','gen'], histtype='step')
+						plt.xlabel('Output of BDT')
+						plt.legend(loc='upper right')
+						plt.savefig('%s%s/BDT_out.png'%(working_directory,saving_directory), bbox_inches='tight')
+						plt.close('all')
 
 					ROC_AUC_SCORE_curr = roc_auc_score(np.append(np.ones(np.shape(out_real[:,1])),np.zeros(np.shape(out_fake[:,1]))),np.append(out_real[:,1],out_fake[:,1]))
 
@@ -544,21 +567,24 @@ for epoch in range(int(1E30)):
 
 					if ROC_AUC_SCORE_list[-1][1] < best_ROC_AUC:
 						print('Saving best ROC_AUC.')
-						generator.save('%s%s/Generator_best_ROC_AUC.h5'%(working_directory,saving_directory))
-						discriminator.save('%s%s/Discriminator_best_ROC_AUC.h5'%(working_directory,saving_directory))
+						if mode != "ROC_testing":
+							generator.save('%s%s/Generator_best_ROC_AUC.h5'%(working_directory,saving_directory))
+							discriminator.save('%s%s/Discriminator_best_ROC_AUC.h5'%(working_directory,saving_directory))
 						# discriminator.save_weights('%s%s/Discriminator_best_ROC_AUC_weights.h5'%(working_directory,saving_directory))
 						best_ROC_AUC = ROC_AUC_SCORE_list[-1][1]
-						shutil.copy('%s%s/CORRELATIONS.png'%(working_directory,saving_directory), '%s%s/BEST_ROC_AUC_Correlations.png'%(working_directory,saving_directory))
+						if mode != "ROC_testing":
+							shutil.copy('%s%s/CORRELATIONS.png'%(working_directory,saving_directory), '%s%s/BEST_ROC_AUC_Correlations.png'%(working_directory,saving_directory))
 
-					plt.figure(figsize=(8,4))
-					plt.title('ROC_AUC_SCORE_list best: %.4f at %d'%(best_ROC_AUC,ROC_AUC_SCORE_list[np.where(ROC_AUC_SCORE_list==best_ROC_AUC)[0][0]][0]))
-					plt.plot(ROC_AUC_SCORE_list[:,0],ROC_AUC_SCORE_list[:,1])
-					plt.axhline(y=best_ROC_AUC,c='k',linestyle='--')
-					plt.axvline(x=ROC_AUC_SCORE_list[np.where(ROC_AUC_SCORE_list==best_ROC_AUC)[0][0]][0],c='k',linestyle='--')
-					plt.savefig('%s%s/ROC_progress.png'%(working_directory,saving_directory),bbox_inches='tight')
-					plt.close('all')
+					if mode != "ROC_testing":
+						plt.figure(figsize=(8,4))
+						plt.title('ROC_AUC_SCORE_list best: %.4f at %d'%(best_ROC_AUC,ROC_AUC_SCORE_list[np.where(ROC_AUC_SCORE_list==best_ROC_AUC)[0][0]][0]))
+						plt.plot(ROC_AUC_SCORE_list[:,0],ROC_AUC_SCORE_list[:,1])
+						plt.axhline(y=best_ROC_AUC,c='k',linestyle='--')
+						plt.axvline(x=ROC_AUC_SCORE_list[np.where(ROC_AUC_SCORE_list==best_ROC_AUC)[0][0]][0],c='k',linestyle='--')
+						plt.savefig('%s%s/ROC_progress.png'%(working_directory,saving_directory),bbox_inches='tight')
+						plt.close('all')
 
-					np.save('%s%s/FoM_ROC_AUC_SCORE_list'%(working_directory,saving_directory),ROC_AUC_SCORE_list)
+					np.save('%s%s/FoM_ROC_AUC_SCORE_list_%d'%(working_directory,saving_directory,FoM_ID),ROC_AUC_SCORE_list)
 
 				print('Saving complete.')
 				t0 = time.time()
